pizzashop/requirements.txt
Django==3.2.1
mysqlclient==1.4.6
djangorestframework==3.12.2
drf-yasg==1.20.0